watcher/requirements.txt
paddleocr==2.7.3
paddlepaddle==2.6.1
Pillow>=10.0.0
numpy>=1.24,<2.0
